+++ CHANGES.rst
@@ -17,6 +17,9 @@ Unreleased
   for symmetry with :meth:`~Reader.search_entries`.
 * Add a ``feed`` argument to :meth:`~Reader.get_feeds`.
 * Add a ``key`` argument to :meth:`~Reader.get_feed_metadata`.
+* Require at least `requests`_ 2.18 (before, the version was unspecified).
+
+.. _requests: https://requests.readthedocs.io
 
 
 Version 0.19